requirements.txt
alembic==1.16.2
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Pillow==11.0.0
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
stripe==12.2.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.22.0
gunicorn==23.0.0